+++ app/src/main/java/com/infomaniak/mail/ui/main/search/SearchFolderAdapter.kt
@@ -65,6 +65,7 @@ class SearchFolderAdapter(
                     onClickListener(folder, entryName)
                 }
             }
+
             return this
         }
     }
